Add showcase for Go to declaration in Swift, fix type warnings (int -> Int), add showcase for Swift quickdoc

+++ 01 - Navigation/02 - Navigate to Declaration/Sources/User.h
@@ -1,7 +1,7 @@
 #import <Foundation/Foundation.h>
 
 @interface User : NSObject
-- (instancetype)initWithName:(NSString *)name age:(int)age;
+- (instancetype)initWithName:(NSString *)name age:(NSInteger)age;
 - (void)celebrateBirthday;
 @end
